Clean changes related to test instabilities
