+++ RoaringBitmap/build.gradle.kts
@@ -17,10 +17,4 @@ tasks.test {
     mustRunAfter(tasks.checkstyleMain)
     useJUnitPlatform()
     failFast = true
-    testLogging {
-        // We exclude 'passed' events
-        events( "skipped", "failed")
-        showStackTraces = true
-        exceptionFormat = org.gradle.api.tasks.testing.logging.TestExceptionFormat.FULL
-    }
 }
