+++ RoaringBitmap/src/test/java/org/roaringbitmap/ParallelAggregationTest.java
@@ -17,17 +17,21 @@ public class ParallelAggregationTest {
 
   private static ForkJoinPool POOL;
 
+  private static ForkJoinPool BIG_POOL;
+
   private static ForkJoinPool NO_PARALLELISM_AVAILABLE;
 
   @BeforeAll
   public static void init() {
-    POOL = new ForkJoinPool(8);
+    POOL = new ForkJoinPool(4);
+    BIG_POOL = new ForkJoinPool(32);
     NO_PARALLELISM_AVAILABLE = new ForkJoinPool(1);
   }
 
   @AfterAll
   public static void teardown() {
     POOL.shutdownNow();
+    BIG_POOL.shutdownNow();
     NO_PARALLELISM_AVAILABLE.shutdownNow();
   }
 
@@ -124,6 +128,15 @@ public class ParallelAggregationTest {
             .toArray(RoaringBitmap[]::new);
     assertEquals(FastAggregation.or(input),
             POOL.submit(() -> ParallelAggregation.or(input)).join());
+  }
+
+  @Test
+  public void regressionTest() {
+    RoaringBitmap[] input = IntStream.range(0, 513)
+            .mapToObj(i -> testCase().withBitmapAt(0).withArrayAt(1).withRunAt(2).build())
+            .toArray(RoaringBitmap[]::new);
+    assertEquals(FastAggregation.or(input),
+            BIG_POOL.submit(() -> ParallelAggregation.or(input)).join());
   }
 
   @Test
